+++ main.py
@@ -8,7 +8,7 @@
 # dependencias usadas para este archivo raiz
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from src.routes.crud_routes import crud_routes
+from src.routes.crud_notification_routes import crud_notification_routes
 from src.routes.send_routes import sendemail_routes
 from src.routes.send_dinamyc_routes import send_router
 
@@ -50,4 +50,4 @@ allow_headers=["*"],           # Headers personalizados permitidos
 # Aquí se incluyen las rutas definidas en la carpeta 'routes'.
 app.include_router(sendemail_routes)
 app.include_router(send_router)
-app.include_router(crud_routes)
+app.include_router(crud_notification_routes)
